refactor : 컨트롤러 로직 분리
컨트롤러 로직 분리 했습니다.

issue : #48
