+++ src/main/java/com/lotto/lotto_simulator/controller/LottoController.java
@@ -16,30 +16,6 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api")
 public class LottoController {
     private final LottoService lottoService;
-    private final PurchaseService purchaseService;
-    private final RoundService roundService;
-
-    //로또 더미데이터 여러개
-    @CrossOrigin
-    @PostMapping("/lottoAutos/{nums}")
-    public ResponseDto<?> lottoAutos(@PathVariable Long nums) {
-        return purchaseService.lottoCreates(nums);
-    }
-
-    //당첨 번호 추첨
-    @CrossOrigin
-    @GetMapping("/winningNums/{num}")
-    public ResponseDto<?> lottoWins(@PathVariable Long num) {
-        return roundService.winningNums(num);
-    }
-
-    //로또 수동 구매
-    @CrossOrigin
-    @PostMapping("/lottoManuals")
-    public ResponseDto<?> lottoManual(@RequestBody LottoDto lottoDto) {
-        return purchaseService.lottoManual(lottoDto);
-    }
-
 
     //feature/uniqueCodeSearch
     @CrossOrigin
@@ -47,27 +23,5 @@ public class LottoController {
     public ResponseDto<?> lottoInfo(@PathVariable Long num,
                                     @RequestBody UniqueCodeDto uniqueIdDto) {
         return lottoService.lottoInfo(num, uniqueIdDto);
-    }
-
-    //test
-    //800만개 난수로직 에 추가 한것 테스트용
-    @CrossOrigin
-    @GetMapping("/lottosAutos/{nums}")
-    public ResponseDto<?> lottoAutoss(@PathVariable Long nums) {
-        return purchaseService.lottoCombinationCreate(nums);
-    }
-
-    //800만게 쿼리문으로 렌덤하게 찾아와서 번호 생성 테스트용
-    @CrossOrigin
-    @GetMapping("/lottosAutoss/{num}")
-    public ResponseDto<?> lottoCombinationCreates(@PathVariable Long num) {
-        return purchaseService.lottoCombinationCreates(num);
-    }
-
-    //페이징 기능 달린 로또 회차 당첨 등수.
-    @CrossOrigin
-    @GetMapping("/winningNum")
-    public ResponseDto<?> winningNum(@PageableDefault(page = 0, size = 1) Pageable pageable) {
-        return lottoService.winningNum(pageable);
     }
 }
